IGNITE-16866: Integer value is out of bounds when the dataregion maxSize is large enough

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/evict/Random2LruPageEvictionTracker.java
@@ -69,7 +69,7 @@ public class Random2LruPageEvictionTracker extends PageAbstractEvictionTracker {
 
     /** {@inheritDoc} */
     @Override public void start() throws IgniteException {
-        trackingArrPtr = GridUnsafe.allocateMemory(trackingSize * 8);
+        trackingArrPtr = GridUnsafe.allocateMemory((long)trackingSize * 8);
 
         GridUnsafe.setMemory(trackingArrPtr, trackingSize * 8, (byte)0);
     }
